requirements.txt
numpy
pandas==2.2.3
pandas_gbq==0.27.0
google-cloud-bigquery==3.29.0
requests
selenium